Misc. fixes before merge
* Moved yaml_ordered_load to config.py
* Reverted batch size change in compute summary for now
* Changed fallback when checking "parallel_model" at checkpoint load to True
* Fixed call to YAML load function in lang model sample after name change
