--- examples/quantization/alexnet_bn_wrpn.yaml
@@ -1,37 +1,0 @@
-quantizers:
-  wrpn_quantizer:
-    class: WRPNQuantizer
-    bits_activations: 8
-    bits_weights: 4
-    bits_overrides:
-      features.0:
-        wts: null
-        acts: null
-      features.1:
-        wts: null
-        acts: null
-      classifier.5:
-        wts: null
-        acts: null
-      classifier.6:
-        wts: null
-        acts: null
-
-lr_schedulers:
-  training_lr:
-    class: MultiStepLR
-    milestones: [60, 75]
-    gamma: 0.2
-
-policies:
-    - quantizer:
-        instance_name: wrpn_quantizer
-      starting_epoch: 0
-      ending_epoch: 200
-      frequency: 1
-
-    - lr_scheduler:
-        instance_name: training_lr
-      starting_epoch: 0
-      ending_epoch: 200
-      frequency: 1
